backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate
mysql-connector-python==9.0.0
python-dotenv==1.0.0
python-multipart==0.0.9
pandas==2.1.4
numpy==1.26.3
celery==5.3.4
redis==5.0.1
PyMuPDF==1.24.6
chromadb==0.4.22
torch==2.5.1
sentence-transformers==3.3.1
groq==0.11.0
instructor==1.6.4
pydantic==2.9.2
fuzzywuzzy
python-Levenshtein
gunicorn
python-docx
PyPDF2
Flask-Login==0.6.3
httpx==0.27.2
